Add files via upload
The objective of the assignment is to analyse and create static visualization from the data set drinks.csv. In order to achieve the objective of this work Pandas library will be used to handle the data and seaborn library for to create the visualizations. To analyse and get insight of the top 5 countries which consume the biggest amount of alcohol.